Use namespaced CMake variables. Deprecated old values
This commit deprecates the old CMake variables in favor of variables
that are namespaced with the AF_ prefix to avoid name conflicts with
user code. If the old variables are used a warning is displayed and
the new variable is set to the value. The old variable is unset to
avoid showing the warning again.